Update CoLM.F90 to support multiple MPI sub group
The goal of this modification is to change the MPI operation mode of CoLM from MPI monopolizing the entire task to secondary grouping after entering the program. The purpose of this is to support the original exclusive use of the entire communication space at the same time, and to divide a communication space into multiple groups. The second method is currently to support data exchange between CoLM and other programs through MPI, such as MPI communication between CoLM and python.